requirements.txt
decord==0.6.0
numpy==1.23.1
torch==1.12.0
ultralytics_yolov5==0.1.1
